refactor(api-types): endpoint 별로 파일 분리
