+++ packages/api-types/src/index.ts
@@ -1,2 +1,6 @@
-export * from "./auth";
 export * from "./shared";
+
+export * from "./auth/common";
+export * from "./auth/complete-signup";
+export * from "./auth/send-mail";
+export * from "./auth/verify-mail";
